доп

+++ Programming/LAB7/server/src/commands/collection/Back.java
@@ -1,9 +1,9 @@
 package commands.collection;
 
 import commands.Command;
-import common.general.Response;
-import common.general.ResponseType;
-import common.general.User;
+import common.exceptions.InvalidInputException;
+import common.net.*;
+import common.tools.Validator;
 import dao.DBManager;
 import managers.CollectionManager;
 
@@ -15,22 +15,22 @@ public class Back extends Command {
         super(user);
     }
 
-    public boolean validateParams(Object... params) {
-        if (params.length == 0 || !(params[0] instanceof String n)) return false;
-        int number;
-        try {
-            number = Integer.parseInt(n);
-            if (number < 0) throw new NumberFormatException();
-        } catch (NumberFormatException e) {
-            return false;
-        }
-        return true;
+    public void validateParams(Object... params) {
+        if (params.length == 0 || !(params[0] instanceof String n)) throw new InvalidInputException("n - не положительное целое число.");
+        if (!Validator.isInt(n) || Integer.parseInt(n) <= 0) throw new InvalidInputException("n - не положительное целое число.");
+    }
+
+    public String getRequiredPermission() {
+        return "CREATE_OBJECT";
     }
 
     public Response execute(CollectionManager cm, DBManager db, Object... params) {
         try {
             int n = Integer.parseInt((String)params[0]);
             String message = cm.back(n, db);
+            if (n > cm.getCommandsList().size()) {
+                return new Response(ResponseType.COMMAND_ERROR, "Было выполнено только " + cm.getCommandsList().size() + " команд");
+            }
             for (int i = 0; i < n; i++) {
                 db.deleteHistoryCommand(getUser());
                 cm.getCommandsList().pop();
